refactor: update import paths to use relative paths instead of absolute
fix: update zod import to use zod/v4 in value objects
docs: remove duplicate comment in graphql schema
style: clean up graphql schema by removing duplicate mutation
