+++ src/domains/authentication/aggregates/value-objects/auth/account-type.vo.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export enum AccountTypeEnum {
   TENANT = 'TENANT',
