+++ src/domains/customer/application/commands/create/review/create-customer-review-product.handler.ts
@@ -1,11 +1,11 @@
 import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 import { CreateCustomerReviewProductDto } from './create-customer-review-product.dto';
 import { CustomerReviewProductDTO } from '../../../mappers/review/customer-review-product.dto';
-import { ICustomerRepository } from 'src/domains/customer/aggregates/repositories/customer.interface';
+import { ICustomerRepository } from '../../../../aggregates/repositories/customer.interface';
 import { Inject } from '@nestjs/common';
-import { ICustomerReviewProductRepository } from 'src/domains/customer/aggregates/repositories/customer-review-product.interface';
+import { ICustomerReviewProductRepository } from '../../../../aggregates/repositories/customer-review-product.interface';
 import { Id } from '@shared/value-objects';
-import { Customer } from 'src/domains/customer/aggregates/entities';
+import { Customer } from '../../../../aggregates/entities';
 import { CustomerReviewProductMapper } from '../../../mappers/review/customer-review-product.mapper';
 
 @CommandHandler(CreateCustomerReviewProductDto)
