+++ src/domains/customer/application/commands/create/wish-list/create-wish-list.dto.ts
@@ -1,4 +1,4 @@
-import { IWishListCreated } from 'src/domains/customer/aggregates/entities';
+import { IWishListCreated } from '../../../../aggregates/entities';
 
 export class CreateWishListDto {
   constructor(
