+++ src/domains/customer/application/commands/update/customer/update-customer.handler.ts
@@ -2,10 +2,10 @@ import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 import { UpdateCustomerDto } from './update-customer.dto';
 import { CustomerMapper } from '../../../mappers';
 import { CustomerDTO } from '../../../mappers';
-import { ICustomerRepository } from 'src/domains/customer/aggregates/repositories/customer.interface';
+import { ICustomerRepository } from '../../../../aggregates/repositories/customer.interface';
 import { Id } from '@shared/value-objects';
 import { Inject, NotFoundException } from '@nestjs/common';
-import { Customer } from 'src/domains/customer/aggregates/entities';
+import { Customer } from '../../../../aggregates/entities';
 
 @CommandHandler(UpdateCustomerDto)
 export class UpdateCustomerHandler
