+++ src/domains/customer/application/events/customer/customer-created.handler.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { EventsHandler, IEventHandler } from '@nestjs/cqrs';
-import { CustomerCreatedEvent } from 'src/domains/customer/aggregates/events';
+import { CustomerCreatedEvent } from '../../../aggregates/events';
 
 @Injectable()
 @EventsHandler(CustomerCreatedEvent)
