+++ src/domains/customer/application/events/review/review-deleted.handler.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { EventsHandler, IEventHandler } from '@nestjs/cqrs';
-import { CustomerReviewProductDeletedEvent } from 'src/domains/customer/aggregates/events';
+import { CustomerReviewProductDeletedEvent } from '../../../aggregates/events';
 
 @Injectable()
 @EventsHandler(CustomerReviewProductDeletedEvent)
