+++ src/domains/customer/application/events/wishlist/wish-list-many-items-deleted.handler.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { EventsHandler, IEventHandler } from '@nestjs/cqrs';
-import { WishlistManyItemsDeletedEvent } from 'src/domains/customer/aggregates/events';
+import { WishlistManyItemsDeletedEvent } from '../../../aggregates/events';
 
 @Injectable()
 @EventsHandler(WishlistManyItemsDeletedEvent)
