+++ src/domains/customer/application/queries/one/customer/find-customer-by-id.handler.ts
@@ -1,7 +1,7 @@
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { FindCustomerByIdDto } from './find-customer-by-id.dto';
 import { Inject, NotFoundException } from '@nestjs/common';
-import { ICustomerRepository } from 'src/domains/customer/aggregates/repositories/customer.interface';
+import { ICustomerRepository } from '../../../../aggregates/repositories/customer.interface';
 import { Id } from '@shared/value-objects';
 import { CustomerDTO, CustomerMapper } from '../../../mappers';
