+++ src/domains/product/aggregates/value-objects/product/type.vo.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export enum TypeEnum {
   PHYSICAL = 'PHYSICAL',
