+++ src/domains/product/aggregates/value-objects/variant/condition.vo.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export enum ConditionEnum {
   NEW = 'NEW',
