+++ src/domains/tenant/aggregates/value-objects/tenant/currency.vo.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export enum CurrencyCodes {
   AFN = 'AFN',
