+++ src/infrastructure/graphql/schema.gql
@@ -55,7 +55,6 @@ type AddressType {
 }
 
 """Types of addresses"""
-"""Types of addresses"""
 enum AddressTypeEnum {
   BILLING
   SHIPPING
@@ -517,9 +516,7 @@ type Mutation {
   addItemToCart(input: AddItemToCartInput!): Cart!
   addReviewProduct(input: CreateCustomerReviewProductInput!): CustomerReviewProduct!
   addStockToWarehouse(input: AddStockToWarehouseInput!, reason: String, variantId: ID!, warehouseId: ID!): Warehouse!
-  addVariant(input: CreateVariantInput!): Product!
   addVariantToWishList(input: WishListItemCreateInput!): Wishlist!
-  addStockToWarehouse(input: AddStockToWarehouseInput!, reason: String, variantId: ID!, warehouseId: ID!): Warehouse!
   archiveVariant(id: String!, productId: String!): Product!
   createAddress(input: CreateAddressInput!): AddressType!
   createCategory(input: CreateCategoryInput!): Category!
@@ -801,17 +798,6 @@ input UpdateCustomerReviewProductInput {
 
   """Updated rating from 1.0 to 5.0"""
   ratingCount: Float
-}
-
-input UpdateDimensionInput {
-  height: Float
-  length: Float
-  width: Float
-}
-
-input UpdateInstallmentInput {
-  interestRate: Float
-  months: Int
 }
 
 input UpdateItemQtyInput {
